# PYTHON-5504 Prototype exponential backoff in with_transaction