PYTHON-5504 Prototype exponential backoff in with_transaction.

Using the repro scrip in jira which runs 200 concurrent transactions in 200 threads all updating the same document shows a significant reduction in wasted retry attempts and latency (from p50 to p100). Before this change:
```
$ python3.13t repro-with_transaction-write-conflict-storm.py
Completed 200 transactions in 200 threads in 4.8626720905303955 seconds
Total retry attempts: 8132
avg latency: 3.04s p50: 3.36s p90: 4.59s p99: 4.83s p100: 4.84s
```

After (with 50ms initial backoff, 1000ms max backoff, and full jitter, backoff starting on the _second_ retry attempt):
```
$ python3.13t repro-with_transaction-write-conflict-storm.py
Completed 200 transactions in 200 threads in 4.251200914382935 seconds
Total retry attempts: 1089
avg latency: 1.53s p50: 1.45s p90: 2.77s p99: 3.69s p100: 4.24s
```

Backoff starting on the _first_ retry attempt appears to work even better:
```
$ python3.13t repro-with_transaction-write-conflict-storm.py
Completed 200 transactions in 200 threads in 3.272695779800415 seconds
Total retry attempts: 886
avg latency: 1.33s p50: 1.21s p90: 2.50s p99: 3.22s p100: 3.25s
```

Note I'm using free-threaded mode to make this repro more similar to the behavior of other languages and other deployment types (eg many single threaded clients running on different machines).