@NoahStapp 2025-08-19
Can you add an async version of the benchmark? Having both APIs be tested before merging this into the backpressure branch would be ideal.